+++ app/src/main/java/com/example/capstone_design/FragmentWeight.java
@@ -118,7 +118,7 @@ public class FragmentWeight extends Fragment {
         return view;
     }
 
-    private String print_status(int weight, float fat) {
+    private String print_status(float weight, float fat) {
         String str;
         str = "강아지의 현재 몸무게는 " + weight + "kg 입니다.\n"
                 + "해당 강아지의 권장 몸무게는 " + fat + "kg 입니다.";
